// apps/backend/.medusa/types/remote-query-entry-points.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: Date | string; output: Date | string; }
  JSON: { input: Record<string, unknown>; output: Record<string, unknown>; }
};
// ...
export type Product = {
  __typename?: 'Product';
  id: Scalars['ID']['output'];
  title: Scalars['String']['output'];
  handle: Scalars['String']['output'];
  subtitle?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  is_giftcard: Scalars['Boolean']['output'];
  status: ProductStatus;
  thumbnail?: Maybe<Scalars['String']['output']>;
  width?: Maybe<Scalars['Float']['output']>;
  weight?: Maybe<Scalars['Float']['output']>;
  length?: Maybe<Scalars['Float']['output']>;
  height?: Maybe<Scalars['Float']['output']>;
  origin_country?: Maybe<Scalars['String']['output']>;
  hs_code?: Maybe<Scalars['String']['output']>;
  mid_code?: Maybe<Scalars['String']['output']>;
  material?: Maybe<Scalars['String']['output']>;
  collection?: Maybe<ProductCollection>;
  collection_id?: Maybe<Scalars['String']['output']>;
  categories?: Maybe<Array<Maybe<ProductCategory>>>;
  type?: Maybe<ProductType>;
  type_id?: Maybe<Scalars['String']['output']>;
  tags: Array<ProductTag>;
  variants: Array<ProductVariant>;
  options: Array<ProductOption>;
  images: Array<ProductImage>;
  discountable?: Maybe<Scalars['Boolean']['output']>;
  external_id?: Maybe<Scalars['String']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  metadata?: Maybe<Scalars['JSON']['output']>;
  sales_channels_link?: Maybe<Array<Maybe<LinkProductSalesChannel>>>;
  sales_channels?: Maybe<Array<Maybe<SalesChannel>>>;
  book_link?: Maybe<LinkProductProductBookModuleBook>;
  book?: Maybe<Book>;
};
// ...
export type Book = {
  __typename?: 'Book';
  id: Scalars['ID']['output'];
  audioVersionUrl?: Maybe<Scalars['String']['output']>;
  dimensions?: Maybe<Scalars['String']['output']>;
  ebookVersionEbook?: Maybe<Scalars['String']['output']>;
  isbn?: Maybe<Scalars['String']['output']>;
  pagesCount?: Maybe<Scalars['Int']['output']>;
  publishDate?: Maybe<Scalars['String']['output']>;
  publisher?: Maybe<Scalars['String']['output']>;
  raw_publishDate?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  author_link?: Maybe<LinkBookModuleBookAuthorModuleAuthor>;
  author?: Maybe<Author>;
  language_link?: Maybe<LinkBookModuleBookLanguageModuleLanguage>;
  language?: Maybe<Language>;
  product_link?: Maybe<LinkProductProductBookModuleBook>;
  product?: Maybe<Product>;
  quote_link?: Maybe<LinkQuoteModuleQuoteBookModuleBook>;
  quote?: Maybe<Quote>;
};

export type Author = {
  __typename?: 'Author';
  id: Scalars['ID']['output'];
  handle: Scalars['String']['output'];
  name: Scalars['String']['output'];
  about: Scalars['String']['output'];
  photoUrl: Scalars['String']['output'];
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  book_link?: Maybe<LinkBookModuleBookAuthorModuleAuthor>;
  book?: Maybe<Book>;
  quote_link?: Maybe<LinkQuoteModuleQuoteAuthorModuleAuthor>;
  quote?: Maybe<Quote>;
};

export type Language = {
  __typename?: 'Language';
  id: Scalars['ID']['output'];
  code: Scalars['String']['output'];
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  book_link?: Maybe<LinkBookModuleBookLanguageModuleLanguage>;
  book?: Maybe<Book>;
};

export type Quote = {
  __typename?: 'Quote';
  id: Scalars['ID']['output'];
  content: Scalars['String']['output'];
  details?: Maybe<Scalars['String']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  author_link?: Maybe<LinkQuoteModuleQuoteAuthorModuleAuthor>;
  author?: Maybe<Author>;
  book_link?: Maybe<LinkQuoteModuleQuoteBookModuleBook>;
  book?: Maybe<Book>;
};
// ...
export type LinkBookModuleBookAuthorModuleAuthor = {
  __typename?: 'LinkBookModuleBookAuthorModuleAuthor';
  book_id: Scalars['String']['output'];
  author_id: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkBookModuleBookLanguageModuleLanguage = {
  __typename?: 'LinkBookModuleBookLanguageModuleLanguage';
  book_id: Scalars['String']['output'];
  language_id: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkProductProductBookModuleBook = {
  __typename?: 'LinkProductProductBookModuleBook';
  product_id: Scalars['String']['output'];
  book_id: Scalars['String']['output'];
  product?: Maybe<Product>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkQuoteModuleQuoteAuthorModuleAuthor = {
  __typename?: 'LinkQuoteModuleQuoteAuthorModuleAuthor';
  quote_id: Scalars['String']['output'];
  author_id: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkQuoteModuleQuoteBookModuleBook = {
  __typename?: 'LinkQuoteModuleQuoteBookModuleBook';
  quote_id: Scalars['String']['output'];
  book_id: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};